# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ody Restaurant API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "preparing",
              "ready",
              "completed",
              "cancelled"
            ]
          },
          "orderType": {
            "type": "string",
            "enum": [
              "dine_in",
              "pickup",
              "delivery"
            ]
          },
          "subtotal": {
            "type": "string"
          },
          "tax": {
            "type": "string"
          },
          "total": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "customerId",
          "status",
          "orderType",
          "subtotal",
          "tax",
          "total",
          "notes",
          "createdAt",
          "completedAt",
          "updatedAt"
        ]
      },
      "CreateOrderItem": {
        "type": "object",
        "properties": {
          "menuItemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 99
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "menuItemId",
          "quantity"
        ]
      },
      "CreateOrder": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "orderType": {
            "type": "string",
            "enum": [
              "dine_in",
              "pickup",
              "delivery"
            ]
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItem"
            },
            "minItems": 1
          }
        },
        "required": [
          "customerId",
          "items"
        ]
      },
      "OrderCustomer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "totalOrders": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "totalOrders"
        ]
      },
      "OrderLineMenuItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "imageUrl"
        ]
      },
      "OrderLineItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "menuItemId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer"
          },
          "priceAtTime": {
            "type": "string"
          },
          "subtotal": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "menuItem": {
            "$ref": "#/components/schemas/OrderLineMenuItem"
          }
        },
        "required": [
          "id",
          "menuItemId",
          "quantity",
          "priceAtTime",
          "subtotal",
          "notes",
          "menuItem"
        ]
      },
      "OrderWithDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Order"
          },
          {
            "type": "object",
            "properties": {
              "customer": {
                "$ref": "#/components/schemas/OrderCustomer"
              },
              "orderItems": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/OrderLineItem"
                }
              }
            },
            "required": [
              "customer",
              "orderItems"
            ]
          }
        ]
      },
      "OrderListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderWithDetails"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "UpdateOrderStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "preparing",
              "ready",
              "completed",
              "cancelled"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create a new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List orders with optional filters",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "preparing",
                "ready",
                "completed",
                "cancelled"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "customerId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "total"
              ]
            },
            "required": false,
            "name": "sortBy",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "required": false,
            "name": "sortOrder",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "last_7_days",
                "last_30_days",
                "all"
              ]
            },
            "required": false,
            "name": "dateFilter",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated order types (dine_in, pickup, delivery). Use none for no matches."
            },
            "required": false,
            "name": "orderTypes",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0,
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{id}/status": {
      "patch": {
        "tags": [
          "orders"
        ],
        "summary": "Update order status with validation",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderWithDetails"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}